use sampler

+++ src/cmd/services/m3coordinator/ingest/config.go
@@ -34,7 +34,7 @@ type Configuration struct {
 	WorkerPoolSize int                          `yaml:"workerPoolSize"`
 	OpPool         pool.ObjectPoolConfiguration `yaml:"opPool"`
 	Retry          retry.Configuration          `yaml:"retry"`
-	LogSampleRate  float32                      `yaml:"logSampleRate" validate:"min=0.0,max=1.0"`
+	LogSampleRate  float64                      `yaml:"logSampleRate" validate:"min=0.0,max=1.0"`
 }
 
 // NewIngester creates an ingester with an appender.
